clean old code
