--- decentralized-api/internal/server/models.go
@@ -1,41 +1,0 @@
-package server
-
-import "net/http"
-
-type InferenceTransaction struct {
-	PromptHash           string `json:"promptHash"`
-	PromptPayload        string `json:"promptPayload"`
-	ResponseHash         string `json:"responseHash"`
-	ResponsePayload      string `json:"responsePayload"`
-	PromptTokenCount     uint64 `json:"promptTokenCount"`
-	CompletionTokenCount uint64 `json:"completionTokenCount"`
-	Model                string `json:"model"`
-	Id                   string `json:"id"`
-}
-
-type ChatRequest struct {
-	Body             []byte
-	Request          *http.Request
-	OpenAiRequest    OpenAiRequest
-	AuthKey          string // signature signing inference request
-	PubKey           string // pubkey of participant, who signed inference request
-	Seed             string
-	InferenceId      string
-	RequesterAddress string // address of participant, who signed inference request
-}
-
-type OpenAiRequest struct {
-	Model     string `json:"model"`
-	Seed      int32  `json:"seed"`
-	MaxTokens int32  `json:"max_tokens"`
-}
-
-type ExecutorDestination struct {
-	Url     string `json:"url"`
-	Address string `json:"address"`
-}
-
-type ResponseWithBody struct {
-	Response  *http.Response
-	BodyBytes []byte
-}
