+++ decentralized-api/internal/server/public/entities.go
@@ -115,3 +115,24 @@ type SubmitUnfundedNewParticipantDto struct {
 	PubKey       string   `json:"pub_key"`
 	WorkerKey    string   `json:"worker_key"`
 }
+
+type UnitOfComputePriceProposalDto struct {
+	Price uint64 `json:"price"`
+	Denom string `json:"denom"`
+}
+
+type PricingDto struct {
+	Price  uint64          `json:"unit_of_compute_price"`
+	Models []ModelPriceDto `json:"models"`
+}
+
+type RegisterModelDto struct {
+	Id                     string `json:"id"`
+	UnitsOfComputePerToken uint64 `json:"units_of_compute_per_token"`
+}
+
+type ModelPriceDto struct {
+	Id                     string `json:"id"`
+	UnitsOfComputePerToken uint64 `json:"units_of_compute_per_token"`
+	PricePerToken          uint64 `json:"price_per_token"`
+}
